Combine all parent files into NCL_Plot
