+++ src/geocat/viz/plot_util/contourf/__init__.py
@@ -1,4 +1,4 @@
 from .contourf import *
-from ._add_geo_features import *
-from ._plot_util import *
-from ._set_up_fig import *
+from _add_geo_features import *
+from _plot_util import *
+from _set_up_fig import *
