+++ src/geocat/viz/plot_util/contourf/contourf.py
@@ -3,6 +3,7 @@
 import xarray as xr
 
 from _set_up_fig import _fig_ax
+
 from _plot_util import NCL_Plot
